requirements.txt
selenium==4.25.0
pyyaml==6.0.2
schedule==1.2.2